+++ src/main/kotlin/com/example/potpot/domain/apply/domain/Apply.kt
@@ -10,6 +10,8 @@ import jakarta.persistence.GenerationType
 import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
+import org.hibernate.annotations.OnDelete
+import org.hibernate.annotations.OnDeleteAction
 
 @Entity(name = "tbl_apply")
 data class Apply(
@@ -22,10 +24,12 @@ data class Apply(
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
+    @OnDelete(action = OnDeleteAction.CASCADE)
     val user: User,
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "feed_id", nullable = false)
+    @OnDelete(action = OnDeleteAction.CASCADE)
     val feed: Feed
 ) {
     constructor() : this(
